requirements.txt
# LLM Client
anthropic>=0.42.0

# MCP SDK
mcp>=1.0.0

# Gmail API
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.150.0

# Web Framework
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Utilities
python-dateutil>=2.9.0
pyyaml>=6.0.2
rich>=13.9.0
python-dotenv>=1.0.1
